refactor: extract generic utilities into utils.py modules
Move fetch_json, ver_sort_key, is_date, to_date, and
filter_supported_eol_entries from redhat.py into a new
rhdh_lifecycle/utils.py module. These functions are used by
AKS/EKS/GKE/PG scripts that have no relation to the Red Hat
Product Life Cycles API.

Create rhdh_prow/utils.py with ver_sort_key (subset of lifecycle
utils). Restore rhdh_prow/__init__.py to docstring-only.

Update all imports across lifecycle and prow scripts. redhat.py
re-exports is_date, to_date, ver_sort_key for backward
compatibility.

Update AGENTS.md shared modules sync rule to reference utils.py.
