+++ skills/lifecycle/scripts/rhdh_lifecycle/ocp.py
@@ -12,7 +12,7 @@ Usage:
 
 import re
 
-from rhdh_lifecycle.redhat import is_date, to_date, ver_sort_key
+from rhdh_lifecycle.utils import is_date, to_date, ver_sort_key
 
 
 def classify_ocp_versions(api_data, today):
