+++ skills/lifecycle/scripts/rhdh_lifecycle/redhat.py
@@ -21,6 +21,8 @@ import urllib.error
 import urllib.parse
 import urllib.request
 
+from rhdh_lifecycle.utils import is_date, to_date, ver_sort_key  # noqa: F401
+
 LIFECYCLE_API_URL = "https://access.redhat.com/product-life-cycles/api/v1/products"
 
 PRODUCT_ALIASES = {
@@ -31,20 +33,6 @@ PRODUCT_ALIASES = {
     "rosa": "Red Hat OpenShift Service on AWS",
     "osd": "Red Hat OpenShift Dedicated",
 }
-
-
-def is_date(val):
-    """Return True if val looks like a YYYY-MM-DD date string."""
-    if not val or not isinstance(val, str):
-        return False
-    return bool(re.match(r"^\d{4}-\d{2}-\d{2}", val))
-
-
-def to_date(val):
-    """Extract YYYY-MM-DD from a date string, or None."""
-    if is_date(val):
-        return val[:10]
-    return None
 
 
 def _phase_date(phases, phase_name):
@@ -58,57 +46,9 @@ def _phase_date(phases, phase_name):
     return "N/A"
 
 
-def ver_sort_key(version_str):
-    """Sort key for version strings like '4.16' or '26.2'."""
-    try:
-        return [int(x) for x in version_str.split(".")]
-    except ValueError:
-        return [0]
-
-
-def filter_supported_eol_entries(eol_data, today):
-    """Filter endoflife.date entries to those still supported.
-
-    Considers both ``eol`` and ``extendedSupport`` fields. Returns the
-    filtered list sorted by cycle version (newest first).
-    """
-    supported = []
-    for entry in eol_data:
-        eol = entry.get("eol", "N/A")
-        ext = entry.get("extendedSupport", "N/A")
-        has_support = False
-        if eol == "N/A":
-            has_support = True
-        elif isinstance(eol, bool):
-            has_support = not eol
-        elif isinstance(eol, str) and eol > today:
-            has_support = True
-        if not has_support and isinstance(ext, str) and ext > today:
-            has_support = True
-        if has_support:
-            supported.append(entry)
-    supported.sort(key=lambda e: ver_sort_key(e["cycle"]), reverse=True)
-    return supported
-
-
 def resolve_product_name(product):
     """Resolve a product alias to the full API product name."""
     return PRODUCT_ALIASES.get(product.lower(), product)
-
-
-def fetch_json(url):
-    """Fetch JSON from a URL.
-
-    Returns the parsed JSON, or None on failure. Shared by all lifecycle
-    scripts that consume external APIs (endoflife.date, AKS, EKS, etc.).
-    """
-    req = urllib.request.Request(url, headers={"User-Agent": "rhdh-skill"})
-    try:
-        with urllib.request.urlopen(req, timeout=30) as resp:
-            return json.loads(resp.read().decode("utf-8"))
-    except (urllib.error.URLError, OSError) as exc:
-        print(f"ERROR: Failed to fetch {url}: {exc}", file=sys.stderr)
-        return None
 
 
 def fetch_api(product_name):
