+++ skills/prow/scripts/analyze_coverage.py
@@ -24,8 +24,8 @@ import urllib.request
 from datetime import datetime, timezone
 from pathlib import Path
 
-from rhdh_prow import ver_sort_key
 from rhdh_prow.repo import resolve_repo_root
+from rhdh_prow.utils import ver_sort_key
 from rhdh_prow.yaml import extract_branch, fetch_yaml, list_yaml_files
 
 POOL_DIR = "clusters/hosted-mgmt/hive/pools/rhdh"
